+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 # 4.2.2-SNAPSHOT
 * Fixed `AsyncDrawable` display when it has placeholder with empty bounds ([#189])
+* Fixed `syntax-highlight` where code input is empty string ([#192])
 
 [#189]: https://github.com/noties/Markwon/issues/189
+[#192]: https://github.com/noties/Markwon/issues/192
 
 # 4.2.1
 * Fix SpannableBuilder `subSequence` method